server/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-mysql-ssl
mysqlclient
python-dotenv
six
django-allauth
requests
cryptography
pillow